Update ExcursionFacade.java
createExcursion return id of Excursion = Long

+++ Travel-Agency-api/src/main/java/fi/muni/cz/pa165/travelagency/facade/ExcursionFacade.java
@@ -22,7 +22,7 @@ public interface ExcursionFacade {
      * @param e *CHANGE ME*
      * @return *CHANGE ME*
      */
-    int createExcursion(ExcursionCreateDTO e);
+    Long createExcursion(ExcursionCreateDTO e);
     
     /**
      *
